requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aioredis==2.0.1
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
babel==2.16.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
fluent-compiler==0.3
fluent.syntax==0.19.0
fluentogram==1.1.10
fonttools==4.55.3
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
importlib_resources==6.5.2
kiwisolver==1.4.7
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.0
multidict==6.1.0
nats-py==2.9.0
numpy==2.0.2
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
psycopg2-binary==2.9.9
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-i18n==0.3.9
pytz==2024.2
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.37
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
watchdog==2.3.1
yarl==1.18.3
zipp==3.21.0
